+++ src/main/java/insane96mcp/iguanatweaksreborn/module/stacksize/StackSize.java
@@ -1,7 +1,7 @@
 package insane96mcp.iguanatweaksreborn.module.stacksize;
 
 import insane96mcp.iguanatweaksreborn.module.stacksize.feature.CustomStackSize;
-import insane96mcp.iguanatweaksreborn.module.stacksize.feature.StackReduction;
+import insane96mcp.iguanatweaksreborn.module.stacksize.feature.GeneralStacking;
 import insane96mcp.iguanatweaksreborn.setup.Config;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
@@ -9,13 +9,13 @@ import insane96mcp.insanelib.base.Module;
 @Label(name = "Stack Size")
 public class StackSize extends Module {
 
-	public StackReduction stackReduction;
+	public GeneralStacking generalStacking;
 	public CustomStackSize customStackSize;
 
 	public StackSize() {
 		super(Config.builder);
 		pushConfig(Config.builder);
-		stackReduction = new StackReduction(this);
+		generalStacking = new GeneralStacking(this);
 		customStackSize = new CustomStackSize(this);
 		Config.builder.pop();
 	}
@@ -23,7 +23,7 @@ public class StackSize extends Module {
 	@Override
 	public void loadConfig() {
 		super.loadConfig();
-		stackReduction.loadConfig();
+		generalStacking.loadConfig();
 		customStackSize.loadConfig();
 	}
 }