+++ changelog.md
@@ -6,6 +6,10 @@
   * Shield Slowdown moved from Stats to Weighted Equipment and the slowdown is now configurable
   * Shield slowdown reduced (25% -> 18% slowdown)
   * Reduced armor slowdown 
+* Stack Sizes
+  * General Stacking (renamed from Stack Reduction)
+    * Items and Blocks less stacking is now disabled by default
+    * Stews now stack up to 16 if Food Stack Reduction is disabled, configurable
 * Soups are now drinked instead of eaten
 * Taking damage from non-entities (e.g. falling, poison) will no longer stop eating/drinking